aoc-2022/src/test/kotlin/com/pchrzasz/aoc/Day03Test.kt
package com.pchrzasz.aoc

import org.assertj.core.api.Assertions
import org.junit.jupiter.api.DisplayName
import org.junit.jupiter.api.Nested
import org.junit.jupiter.api.Test

@DisplayName("Day 3")
class Day03Test {

    private val input = """
        vJrwpWtwJgWrhcsFMMfFFhFp
        jqHRNqRjqzjGDLGLrsFMfFZSrLrFZsSL
        PmmdzqPrVvPwwTWBwg
        wMqvLMZHhHMvwLHjbvcjnnSBnvTQFn
        ttgJtRGJQctTZtZT
        CrZsJsPPZsGzwwsLwLmpwMDw
    """.trimIndent()

    @Nested
    @DisplayName("Part 1")
    inner class Part1 {
        @Test
        fun `Matches example`() {
            val answer = Day03().solvePart1(input)

            Assertions.assertThat(answer).isEqualTo(157)
        }

        @Test
        fun `Actual answer`() {
            val inputAsText = Resources.resourceAsText("input_day03.txt")
            val answer = Day03().solvePart1(inputAsText)

            Assertions.assertThat(answer).isEqualTo(7_737)
        }
    }

    @Nested
    @DisplayName("Part 2")
    inner class Part2 {
        @Test
        fun `Matches example`() {
            val answer = Day03().solvePart2(input)

            Assertions.assertThat(answer).isEqualTo(70)
        }

        @Test
        fun `Actual answer`() {
            val inputAsText = Resources.resourceAsText("input_day03.txt")
            val answer = Day03().solvePart2(inputAsText)

            Assertions.assertThat(answer).isEqualTo(2_697)
        }
    }

}